Branch changes to use switch instead of checkout
When using WSL2 with zsh, the `git co` autocomplete for branches is
really slow (like 5s to complete a branch name). This is most likely to
do with certain repos being on an ntfs drive while WSL2 is prob using
something else. I dont know. It's what the internet was suggesting.

By using `git switch` the autocomplete is much faster. This is the new
and preferred way of doing this anyway, so...not really a big problem.